Initial commit
Mappington is a small library that makes generating uris pointing to
Google's static map API a little easier. I intend to only support a
small subset of the API; I won't be implementing feature styles, for
example, as I don't see the use in them. Using this library you'll be
able to:

* Quickly make a map centered on a location using default values.
* Create a map with custom parameters.
* Render markers and paths on a map.