Add Point and PointList
Markers and paths need points. Points can be either a string or
an array. A string should be an address of some kind; an array indicates
a coordinate, and it will be interpreted as such. The point will render
the address or coordinate to a string. You don't need to use points
directly; just pass a string or an array to a marker or path.