Create Mappington::Map and add spec
A map can be centered on a point, can render markers, and you can draw
on the map with paths. You must specify either a center point and a zoom
level, or at least 1 marker or path, or the map will raise an error. You
must also tell the map how big it should be by setting it's size.

Mappington::Map will return a string that points to a map rendered by
Google. The map should always return a valid URI or raise an error.