Let Google know if you use a location sensor
Google demands to know if you are using a location such as GPS. You will
not get a map back without specifying this query parameter. A sensor
variable has been added to Map. The value is false by default, so you
should only need to set it if you use a location sensor.